src/app/requirements.txt
altair==4.0
streamlit==1.12.0
Pillow==9.3.0
dlib==19.24.2